Rename file back to Duke.java and edit some print statements
